compute/pipelines: rename stuff for cleanup
